Fix review comments on app attributes rest api
