+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.store.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/store/v1/ApplicationsApi.java
@@ -248,23 +248,6 @@ ApplicationsApiService delegate = new ApplicationsApiServiceImpl();
     }
 
     @GET
-    @Path("/attributes")
-    @Consumes({ "application/json" })
-    @Produces({ "application/json" })
-    @ApiOperation(value = "Get all application attributes from configuration ", notes = "This operation can be used to retrieve the application attributes from configuration. It will not return hidden attributes. ", response = Void.class, authorizations = {
-        @Authorization(value = "OAuth2Security", scopes = {
-            @AuthorizationScope(scope = "apim:subscribe", description = "Subscribe API")
-        })
-    }, tags={ "Application Attributes",  })
-    @ApiResponses(value = { 
-        @ApiResponse(code = 200, message = "OK. Application attributes returned. ", response = Void.class),
-        @ApiResponse(code = 404, message = "Not Found. Requested application does not exist. ", response = ErrorDTO.class),
-        @ApiResponse(code = 406, message = "Not Acceptable. The requested media type is not supported ", response = ErrorDTO.class) })
-    public Response applicationsAttributesGet(@ApiParam(value = "Validator for conditional requests; based on the ETag of the formerly retrieved variant of the resourec. " )@HeaderParam("If-None-Match") String ifNoneMatch) {
-        return delegate.applicationsAttributesGet(ifNoneMatch, securityContext);
-    }
-
-    @GET
     
     @Consumes({ "application/json" })
     @Produces({ "application/json" })
