+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store-new/source/src/app/data/api.jsx
@@ -231,9 +231,14 @@ export default class API extends Resource {
         }
     }
 
+    /**
+     * Get all application attributes
+     * @param {function} callback which needs to be called upon success
+     * @returns {promise} With given callback attached to the success chain else API invoke promise.
+     */
     getAllApplicationAttributes() {
         return this.client.then((client) => {
-            return client.apis['Application Attributes'].get_applications_attributes(this._requestMetaData());
+            return client.apis['Application Attributes'].get_settings_attributes(this._requestMetaData());
         });
     }
